+++ android/src/main/java/com/cloudwebrtc/webrtc/FlutterWebRTCPlugin.java
@@ -7,6 +7,7 @@ import android.os.Bundle;
 import android.util.Log;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.lifecycle.DefaultLifecycleObserver;
 import androidx.lifecycle.Lifecycle;
 import androidx.lifecycle.LifecycleOwner;
@@ -17,8 +18,10 @@ import com.cloudwebrtc.webrtc.utils.AnyThreadSink;
 import com.cloudwebrtc.webrtc.utils.ConstraintsMap;
 
 import org.webrtc.ExternalAudioProcessingFactory;
+import org.webrtc.MediaStream;
 import org.webrtc.MediaStreamTrack;
 import org.webrtc.PeerConnectionFactory;
+import org.webrtc.audio.JavaAudioDeviceModule;
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
@@ -62,6 +65,10 @@ public class FlutterWebRTCPlugin implements FlutterPlugin, ActivityAware, EventC
         return methodCallHandler.getTrackForId(trackId, peerConnectionId);
     }
 
+    public MediaStream getStreamForId(String streamId, String peerConnectionId) {
+        return methodCallHandler.getStreamForId(streamId, peerConnectionId);
+    }
+
     public LocalTrack getLocalTrack(String trackId) {
         return methodCallHandler.getLocalTrack(trackId);
     }
@@ -72,6 +79,11 @@ public class FlutterWebRTCPlugin implements FlutterPlugin, ActivityAware, EventC
 
     public PeerConnectionFactory getPeerConnectionFactory() {
         return methodCallHandler.getPeerConnectionFactory();
+    }
+
+    @Nullable
+    public JavaAudioDeviceModule getAudioDeviceModule() {
+        return methodCallHandler == null ? null : methodCallHandler.getAudioDeviceModule();
     }
 
     @Override